Enable 64-bit build in Linux and Windows - Update CMakeLists.txt to allow 64-bit building - Add lib64 option for running 64-bit Dr.Memory - Add many type cast - Disable some syscall handling - Config only two tests for 64-bit pattern mode. - Replace regs in _CONTEXT - add ASSERT_TRUNCATE
SVN-Revision: 871

+++ tests/malloc.c
@@ -207,8 +207,8 @@ main()
         static char *p;
         p = malloc(8); /* static so no leak */
         free(p);
-        if (0xfff - ((int)p & 0xfff) < 512) /* truncation ok in cast */
-            p = malloc(0xfff - ((int)p & 0xfff));
+        if (0xfff - ((int)(size_t)p & 0xfff) < 512) /* truncation ok in cast */
+            p = malloc(0xfff - ((int)(size_t)p & 0xfff));
         else
             p = NULL; /* prevent from pointing where p1 will be (i#340) */
     }
